Integration2/requirements.txt
networkx==3.2
pandas==2.1.1
Pillow==10.2.0
opencv-contrib-python==4.8.1.78
ultralytics==8.0.200
streamlit==1.29.0
ttkbootstrap==1.10.1
twilio==8.13.0
torch==2.1.0
torchaudio==2.1.2
torchvision==0.16.0
tensorflow==2.15.0
numpy==1.26.2
keras==2.15.0